Move fetching logic away
